api/requirements.txt
django<=5.0.6
djongo==1.3.6
bcrypt==4.1.2
requests==2.31.0
PyJWT==2.8.0
pymongo==3.12.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Fuzzy==1.2.2
